% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genero.R
\name{genero}
\alias{genero}
\title{Panel component for shiny panels layout}
\usage{
genero(
  names,
  result_as = c(male = "male", female = "female"),
  lang = "es",
  col = NULL,
  na = NA,
  rev_weights = FALSE
)
}
\arguments{
\item{names}{A vector or data.frame with names or full names}

\item{result_as}{A named vector with names c("male", "female")
values can be used to override the results}

\item{lang}{Use "es" for Spanish (default), "pt" for Portuguese.}

\item{col}{The name of the column with the names or full names
when the input is a data frame.}

\item{na}{String to be used when there is not match for gender}

\item{rev_weights}{Boolean to indicate if weights should be
reversed when input names have the format Last Name First Name}
}
\value{
A vector of data frame with the estimated gender for the
  input. When the input is data.frame a column is attached next
  to the column used for the input names with the result.
}
\description{
Panel component for shiny panels layout
}
\examples{
genero(c("Juan", "Pablo", "Camila", "Mariana"))


}
