% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-names_gender_es.R
\docType{data}
\name{names_gender_es}
\alias{names_gender_es}
\title{Names with gender in Spanish}
\format{Data frame with two columns: name and gender.}
\usage{
names_gender_es
}
\description{
These data was collected and organized manually from multiples sources.
It consists of more than 9810 names in Spanish and its corresponding associated
gender accounting for name variations.
}
\examples{
names_gender_es
}
\keyword{datasets}
