% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-names_gender_pt.R
\docType{data}
\name{names_gender_pt}
\alias{names_gender_pt}
\title{Names with gender in Portuguese}
\format{Data frame with two columns: name and gender.}
\usage{
names_gender_pt
}
\description{
These data is created and derived from https://brasil.io/dataset/genero-nomes/nomes
it consists of more than 50.000 names in Portuguese and its corresponding associated
gender.
}
\examples{
names_gender_pt
}
\keyword{datasets}
