% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genero.R
\name{which_name_column}
\alias{which_name_column}
\title{Which name column}
\usage{
which_name_column(colnames, colname_variations = NULL, show_guess = FALSE)
}
\arguments{
\item{colnames}{A vector of data.frame names.}

\item{colname_variations}{A vector of custom names to append to the vector
of frequent colnames for first names.}

\item{show_guess}{Show message with the guessed column.}
}
\value{
A single colname with the match of common first name columns.
}
\description{
Which name column
}
\examples{
which_name_column(c("Name", "Age", "City"))

}
